+++ lib/salve/validate.ts
@@ -8,8 +8,8 @@
 
 export const version: string = "4.0.3-rc1";
 
-export { eventsToTreeString, Event, Grammar, GrammarWalker, BasePattern,
-         RefError, Walker, __test, FireEventResult,
+export { eventsToTreeString, Event, EventSet, Grammar, GrammarWalker,
+         BasePattern, RefError, Walker, __test, FireEventResult,
          EndResult } from "./patterns";
 
 export { constructTree } from "./formats";