+++ CHANGELOG.rst
@@ -3,6 +3,10 @@ patch part of the version number (i.e. the number after the 2nd dot)
 are generally not listed here unless they include a fix to a specific
 issue reported on github.
 
+* 4.0.4:
+
+  - Allow passing an object to ``constructTree``.
+
 * 4.0.3:
 
   - Fixed an obscure bug revealed through linting the code more stringently.